+++ R/shiny.R
@@ -1158,7 +1158,7 @@ ShinySession <- R6Class(
         attr(label, "srcfile") <- srcfile
 
         # Do not bind this `observe()` call
-        obs <- with_no_otel_bind(observe(..stacktraceon = FALSE, {
+        obs <- with_no_otel_collect(observe(..stacktraceon = FALSE, {
 
           private$sendMessage(recalculating = list(
             name = name, status = 'recalculating'